+++ assets/js/jquery.smooth-scroll.js
@@ -1,0 +1,42 @@
+/**
+ * Fulcrum Site JavaScript handling.  This script handles opening/closing of the questions and answers,
+ * toggling of the icon handle, and setting of the class states.
+ *
+ * @package     Fulcrum Site
+ * @since       1.0.0
+ * @author      hellofromTonya
+ * @link        http://hellofromtonya.com
+ * @license     GNU General Public License 2.0+
+ */
+;(function ( $, window, document, undefined ) {
+	'use strict'
+
+	var $body;
+
+	function init() {
+		$body = $( 'html, body' );
+
+		$( 'a[href^="#"]' ).on( 'click', function( event ) {
+			event.preventDefault();
+
+			smoothScrollHandler( this.hash, event );
+		});
+	}
+
+	function smoothScrollHandler( target, event ) {
+		var $target = $( target );
+
+		if ( typeof $target == "undefined" || ! $target.length ) {
+			return false;
+		}
+
+		$body.stop().animate({
+			'scrollTop': $target.offset().top - 150
+		}, 2000, 'swing' );
+	}
+
+	$( document ).ready( function () {
+		init();
+	} );
+
+}( jQuery, window, document ));