+++ src/class-plugin.php
@@ -20,7 +20,7 @@ class Plugin extends Addon {
 	 *
 	 * @var string
 	 */
-	const VERSION = '1.0.0';
+	const VERSION = '1.0.1';
 
 	/**
 	 * The plugin's minimum WordPress requirement