+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+- removed scrollup from script
+- optimized smooth scroll script
+
 ## 1.0.0
 
 Initial release.